add d11 p2
Instead of storing the item, we store the moduli of it against the
dividers.  We accomplish it by first, storing everything, including the
original value, into a HashMap where the key is the divider and the
value is the modulus.  The original value would have '0' as it's
divider.  Then after having constructed the initial hashmap, we remove
key '0' and the attached original value.

  ```
  From
  7

  To
  {
    0: 7         // to be removed
    3: 7 % 3 = 1
    5: 7 % 5 = 2
  }
  ```

Then, in subsequent transfers of items between monkey, we simply apply
each monkey's operation to all the value and then store the modulus of
the operation's result against the divider key.

  ```
  From in monkey with operation |x| x + 1
  {
    3: 7 % 3 = 1
    5: 7 % 5 = 2
  }

  To
  {
    3: (2 + 1) % 3 = 0
    5: (2 + 1) % 5 = 3
  }
  ```

This is because we could take advantage of the following:

  ```
  (a + b) % c == ((a % c) + (b % c)) % c

  and

  (a * b) % c == ((a % c) * (b % c)) % c
  ```

This helps us reduced the size of the item's value over multiple
iterations.